Added vue-routing for pages. Added simple 'Home', 'Host', and 'Join' pages
