+++ scribblersGame/package.json
@@ -1,32 +1,35 @@
 {
-  "name": "scribblersgame",
+  "name": "vueenvcomparison",
   "version": "0.0.0",
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "vite --host",
-    "build": "vite build",
+    "dev": "vite",
+    "build": "run-p type-check \"build-only {@}\" --",
     "preview": "vite preview",
-    "test:unit": "vitest",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build",
     "lint": "eslint . --fix"
   },
   "dependencies": {
     "pinia": "^3.0.1",
-    "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.20.0",
+    "@tsconfig/node22": "^22.0.0",
+    "@types/node": "^22.13.9",
     "@vitejs/plugin-vue": "^5.2.1",
-    "@vitest/eslint-plugin": "1.1.31",
-    "@vue/test-utils": "^2.4.6",
-    "eslint": "^9.20.1",
-    "eslint-plugin-vue": "^9.32.0",
-    "jsdom": "^26.0.0",
-    "vite": "^6.1.0",
+    "@vue/eslint-config-typescript": "^14.5.0",
+    "@vue/tsconfig": "^0.7.0",
+    "eslint": "^9.21.0",
+    "eslint-plugin-vue": "~10.0.0",
+    "jiti": "^2.4.2",
+    "npm-run-all2": "^7.0.2",
+    "typescript": "~5.8.0",
+    "vite": "^6.2.1",
     "vite-plugin-vue-devtools": "^7.7.2",
-    "vitest": "^3.0.5"
+    "vue-tsc": "^2.2.8"
   }
 }
