+++ scribblersGame/src/components/WelcomeItem.vue
@@ -30,6 +30,7 @@ i {
   place-content: center;
   width: 32px;
   height: 32px;
+
   color: var(--color-text);
 }
