+++ scribblersGame/src/main.ts
