+++ scribblersGame/src/views/HostLobbyView.vue
@@ -1,0 +1,7 @@
+<!-- Simple join screen to display a link that routes to the select avatar screen -->
+<template>
+    <h1>
+        Host screen to be implemented here
+    </h1>
+    <RouterLink to="/avatar">Select Avatar</RouterLink>
+</template>
