+++ scribblersGame/vite.config.ts
